ClipFuse/requirements.txt
requests
python-dotenv
pyht==0.0.28
moviepy==1.0.3
#pillow==9.4.0
ffmpeg==1.4
SpeechRecognition==3.10.4
pydub==0.25.1
langchain-openai==0.1.14
flask==2.2.2